synology-deploy/requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.5
python-dotenv>=1.0.1
python-dateutil>=2.9.0.post0
pytz>=2024.1